feat: move Memory Preferences and Maintenance Autopilot to their main pages, improve Triggers setup UX
- Extract MemoryPreferences to components/settings/MemoryPreferences.tsx
  and render it at the top of the Memory page — preferences now live
  alongside the memories they configure
- Extract MaintenanceSettings to components/settings/MaintenanceSettings.tsx
  and add a new Configuration tab (first/default) to the Maintenance page —
  autopilot setup is now co-located with policies, scans, and audit
- Remove both SettingsSection blocks from Settings page (no longer needed
  there; pages link directly)
- Replace minimal Triggers disabled-state with a full step-by-step setup
  guide: generate key with openssl, add TRIGGER_SECRET_KEY to .env, restart
