+++ client/src/pages/Settings.tsx
@@ -22,12 +22,9 @@ import {
   Save,
   Eye,
   EyeOff,
-  Brain,
   Wrench,
   Terminal,
   Link2,
-  Shield,
-  ShieldCheck,
   Info,
 } from "lucide-react";
 import {
@@ -71,278 +68,6 @@ const CLOUD_PROVIDERS: Array<{
   { key: "google",    name: "Google (Gemini)",    envVar: "GOOGLE_API_KEY"    },
   { key: "xai",       name: "xAI (Grok)",         envVar: "XAI_API_KEY"       },
 ];
-
-const PREFERENCE_ROWS = [
-  { key: "preferred-language", label: "Preferred Language", placeholder: "e.g. TypeScript" },
-  { key: "error-handling-style", label: "Error Handling Style", placeholder: "e.g. throw exceptions, return Result types" },
-  { key: "preferred-db", label: "Preferred Database", placeholder: "e.g. PostgreSQL, SQLite" },
-  { key: "code-style", label: "Code Style", placeholder: "e.g. functional, OOP" },
-  { key: "test-framework", label: "Test Framework", placeholder: "e.g. Jest, Vitest" },
-];
-
-
-// ─── Maintenance Settings ─────────────────────────────────────────────────────
-
-const SCHEDULE_PRESETS = [
-  { label: "Daily at 9am", value: "0 9 * * *" },
-  { label: "Weekly (Mon 9am)", value: "0 9 * * 1" },
-  { label: "Monthly (1st, 9am)", value: "0 9 1 * *" },
-];
-
-const SEVERITY_OPTIONS = ["critical", "high", "medium", "low", "info"] as const;
-
-interface MaintenancePolicySummary {
-  id: string;
-  enabled: boolean;
-  schedule: string;
-  severityThreshold: string;
-}
-
-interface MaintenanceSettingsProps {
-  /** When true, renders content only without a Card wrapper. Use inside SettingsSection. */
-  noCard?: boolean;
-}
-
-function MaintenanceSettings({ noCard = false }: MaintenanceSettingsProps) {
-  const qc = useQueryClient();
-
-  const { data: policies, isLoading } = useQuery<MaintenancePolicySummary[]>({
-    queryKey: ["/api/maintenance/policies"],
-    queryFn: () =>
-      apiRequest("GET", "/api/maintenance/policies") as Promise<MaintenancePolicySummary[]>,
-  });
-
-  const firstPolicy = policies?.[0] ?? null;
-
-  const [schedule, setSchedule] = useState(SCHEDULE_PRESETS[1].value);
-  const [severity, setSeverity] = useState<string>("high");
-
-  const createPolicy = useMutation({
-    mutationFn: () =>
-      apiRequest("POST", "/api/maintenance/policies", {
-        enabled: true,
-        schedule,
-        severityThreshold: severity,
-        categories: [],
-        autoMerge: false,
-        notifyChannels: [],
-      }),
-    onSuccess: () => void qc.invalidateQueries({ queryKey: ["/api/maintenance/policies"] }),
-  });
-
-  const togglePolicy = useMutation({
-    mutationFn: ({ id, enabled }: { id: string; enabled: boolean }) =>
-      apiRequest("PUT", `/api/maintenance/policies/${id}`, { enabled }),
-    onSuccess: () => void qc.invalidateQueries({ queryKey: ["/api/maintenance/policies"] }),
-  });
-
-  const updatePolicy = useMutation({
-    mutationFn: ({ id, schedule: s, severityThreshold }: { id: string; schedule: string; severityThreshold: string }) =>
-      apiRequest("PUT", `/api/maintenance/policies/${id}`, { schedule: s, severityThreshold }),
-    onSuccess: () => void qc.invalidateQueries({ queryKey: ["/api/maintenance/policies"] }),
-  });
-
-  const isEnabled = firstPolicy?.enabled ?? false;
-
-  const inner = (
-    <div className={cn("space-y-4", noCard ? "p-4" : "")}>
-      <p className="text-xs text-muted-foreground">
-        Automatically scan workspaces for dependency updates, security advisories, and license issues.
-        Configure and manage scans in detail from the{" "}
-        <a href="/maintenance" className="text-primary underline underline-offset-2">
-          Maintenance
-        </a>{" "}
-        page.
-      </p>
-
-      {isLoading ? (
-        <div className="flex items-center gap-2 text-xs text-muted-foreground">
-          <Loader2 className="h-3 w-3 animate-spin" />
-          Loading...
-        </div>
-      ) : (
-        <div className="space-y-4">
-          {/* Enable/disable toggle */}
-          <div className="flex items-center justify-between">
-            <div>
-              <p className="text-sm font-medium">Enable Autopilot</p>
-              <p className="text-xs text-muted-foreground">
-                {firstPolicy
-                  ? isEnabled
-                    ? "Autopilot is active — scans run on schedule"
-                    : "Autopilot is paused"
-                  : "No policy configured yet"}
-              </p>
-            </div>
-            <Switch
-              checked={isEnabled}
-              onCheckedChange={(checked) => {
-                if (firstPolicy) {
-                  togglePolicy.mutate({ id: firstPolicy.id, enabled: checked });
-                } else if (checked) {
-                  createPolicy.mutate();
-                }
-              }}
-              disabled={togglePolicy.isPending || createPolicy.isPending}
-            />
-          </div>
-
-          {/* Schedule selector */}
-          <div className="flex items-center gap-3">
-            <label className="text-xs font-medium w-32 shrink-0">Default Schedule</label>
-            <Select
-              value={firstPolicy?.schedule ?? schedule}
-              onValueChange={(val) => {
-                setSchedule(val);
-                if (firstPolicy) {
-                  updatePolicy.mutate({ id: firstPolicy.id, schedule: val, severityThreshold: firstPolicy.severityThreshold });
-                }
-              }}
-            >
-              <SelectTrigger className="h-8 text-xs flex-1">
-                <SelectValue placeholder="Select schedule" />
-              </SelectTrigger>
-              <SelectContent>
-                {SCHEDULE_PRESETS.map((p) => (
-                  <SelectItem key={p.value} value={p.value} className="text-xs">
-                    {p.label}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-
-          {/* Severity threshold selector */}
-          <div className="flex items-center gap-3">
-            <label className="text-xs font-medium w-32 shrink-0">Severity Threshold</label>
-            <Select
-              value={firstPolicy?.severityThreshold ?? severity}
-              onValueChange={(val) => {
-                setSeverity(val);
-                if (firstPolicy) {
-                  updatePolicy.mutate({ id: firstPolicy.id, schedule: firstPolicy.schedule, severityThreshold: val });
-                }
-              }}
-            >
-              <SelectTrigger className="h-8 text-xs flex-1">
-                <SelectValue placeholder="Select threshold" />
-              </SelectTrigger>
-              <SelectContent>
-                {SEVERITY_OPTIONS.map((s) => (
-                  <SelectItem key={s} value={s} className="text-xs capitalize">
-                    {s}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-
-          {(togglePolicy.isError || createPolicy.isError || updatePolicy.isError) && (
-            <p className="text-xs text-destructive">
-              {((togglePolicy.error ?? createPolicy.error ?? updatePolicy.error) as Error).message}
-            </p>
-          )}
-        </div>
-      )}
-    </div>
-  );
-
-  if (noCard) return inner;
-
-  // Legacy standalone card render (backward-compat for any direct usage)
-  return (
-    <div className="rounded-lg border border-border bg-card shadow-sm">
-      <div className="px-4 py-3 border-b border-border">
-        <p className="text-base font-semibold flex items-center gap-2">
-          {isEnabled ? (
-            <ShieldCheck className="h-4 w-4 text-green-500" />
-          ) : (
-            <Shield className="h-4 w-4 text-muted-foreground" />
-          )}
-          Maintenance Autopilot
-        </p>
-      </div>
-      {inner}
-    </div>
-  );
-}
-
-interface MemoryPreferencesProps {
-  /** When true, renders content only without a Card wrapper. Use inside SettingsSection. */
-  noCard?: boolean;
-}
-
-function MemoryPreferences({ noCard = false }: MemoryPreferencesProps) {
-  const qc = useQueryClient();
-  const [values, setValues] = useState<Record<string, string>>({});
-  const [saved, setSaved] = useState<Record<string, boolean>>({});
-
-  const savePreference = useMutation({
-    mutationFn: ({ key, value }: { key: string; value: string }) =>
-      apiRequest("POST", "/api/memories", {
-        scope: "global",
-        type: "preference",
-        key,
-        content: value,
-        confidence: 1.0,
-      }),
-    onSuccess: (_data, { key }) => {
-      setSaved((prev) => ({ ...prev, [key]: true }));
-      void qc.invalidateQueries({ queryKey: ["/api/memories"] });
-      setTimeout(() => setSaved((prev) => ({ ...prev, [key]: false })), 2000);
-    },
-  });
-
-  const inner = (
-    <div className={cn("space-y-3", noCard ? "p-4" : "")}>
-      <p className="text-xs text-muted-foreground">
-        These preferences are stored as global memories and injected into every pipeline stage, helping the AI make consistent decisions aligned with your preferences.
-      </p>
-      {PREFERENCE_ROWS.map(({ key, label, placeholder }) => (
-        <div key={key} className="flex items-center gap-3">
-          <label className="text-xs font-medium w-44 shrink-0">{label}</label>
-          <Input
-            className="h-8 text-xs flex-1"
-            placeholder={placeholder}
-            value={values[key] ?? ""}
-            onChange={(e) => setValues((prev) => ({ ...prev, [key]: e.target.value }))}
-          />
-          <Button
-            size="sm"
-            className="h-8 text-xs shrink-0"
-            disabled={!values[key]?.trim() || savePreference.isPending}
-            onClick={() => {
-              const value = values[key]?.trim();
-              if (value) savePreference.mutate({ key, value });
-            }}
-          >
-            {saved[key] ? (
-              <CheckCircle2 className="h-3 w-3 text-emerald-500" />
-            ) : savePreference.isPending ? (
-              <Loader2 className="h-3 w-3 animate-spin" />
-            ) : (
-              <><Save className="h-3 w-3 mr-1" /> Save</>
-            )}
-          </Button>
-        </div>
-      ))}
-    </div>
-  );
-
-  if (noCard) return inner;
-
-  // Legacy standalone card render
-  return (
-    <div className="rounded-lg border border-border bg-card shadow-sm">
-      <div className="px-4 py-3 border-b border-border">
-        <p className="text-base font-semibold flex items-center gap-2">
-          <Brain className="h-4 w-4" /> Project Memory Preferences
-        </p>
-      </div>
-      {inner}
-    </div>
-  );
-}
 
 export default function Settings() {
   const { data: models, isLoading: modelsLoading } = useModels();
@@ -1385,28 +1110,6 @@ export default function Settings() {
             defaultOpen={false}
           >
             <ArgocdSettings noCard />
-          </SettingsSection>
-
-          {/* ── 9. Maintenance Autopilot ───────────────────── */}
-          <SettingsSection
-            title="Maintenance Autopilot"
-            icon={<Shield className="h-4 w-4" />}
-            shortDescription="Scheduled scanning for dependency updates and security advisories."
-            longDescription="Runs automated scans on a cron schedule to detect outdated dependencies, CVEs, license issues, and config drift. Findings appear in the Maintenance page. Scans use the Development team's configured model."
-            defaultOpen={false}
-          >
-            <MaintenanceSettings noCard />
-          </SettingsSection>
-
-          {/* ── 10. Memory Preferences ─────────────────────── */}
-          <SettingsSection
-            title="Memory Preferences"
-            icon={<Brain className="h-4 w-4" />}
-            shortDescription="Persistent AI preferences injected into pipeline stage prompts."
-            longDescription="These preferences are injected into every pipeline stage's system prompt as context. Use them to encode your team's standards, preferred technologies, and coding style so agents consistently produce on-brand output."
-            defaultOpen={false}
-          >
-            <MemoryPreferences noCard />
           </SettingsSection>
 
           {/* ── 11. Version Information ────────────────── */}
